added facebook developer comment and like plugin code in post details. now further will add draft feature in posts

+++ src/posts/views.py
@@ -1,5 +1,6 @@
+from urllib.parse import quote, quote_plus
 from django.contrib import messages
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.core.paginator import Paginator
 from django.shortcuts import render, get_object_or_404, redirect
 from .models import Post
@@ -33,10 +34,15 @@ from .forms import PostForm
 
 
 def post_create(request):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    if not request.user.is_authenticated():
+        raise Http404
     form=PostForm(request.POST or None, request.FILES or None)
     if request.method == 'POST':
         if form.is_valid():
             instance = form.save(commit=False)
+            instance.user=request.user
             print(form.cleaned_data)
             instance.save()
             messages.success(request,"Successfully Created !!")
@@ -57,7 +63,8 @@ def post_detail(request, slug=None):
    # thats why you have to use get objrect or 404 method.
 
    instance=get_object_or_404(Post,slug=slug)
-   context = {'title': ' detail page', 'instance' : instance}
+   share_string=quote_plus(instance.content)
+   context = {'title': ' detail page', 'instance' : instance, 'share_string':share_string}
    return render(request, 'post_detail.html', context)
 
 def post_list(request):
@@ -80,18 +87,22 @@ def post_list(request):
 
 
 def post_update(request, slug=None):
-   instance=get_object_or_404(Post,slug=slug)
-   form=PostForm(request.POST or None, request.FILES or None, instance=instance)
-   if form.is_valid():
-       instance = form.save(commit=False)
-       print(form.cleaned_data.get('title'))
-       instance.save()
-       messages.success(request,"<a href='#'>Successfully</a> Updated and Saved !!",extra_tags='html_safe')
-       return HttpResponseRedirect(instance.get_absolute_url())
-   context = {'title' : instance.title, 'instance' : instance, 'form' : form}
-   return render(request, 'post_form.html', context)
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    instance=get_object_or_404(Post,slug=slug)
+    form=PostForm(request.POST or None, request.FILES or None, instance=instance)
+    if form.is_valid():
+        instance = form.save(commit=False)
+        print(form.cleaned_data.get('title'))
+        instance.save()
+        messages.success(request,"<a href='#'>Successfully</a> Updated and Saved !!",extra_tags='html_safe')
+        return HttpResponseRedirect(instance.get_absolute_url())
+    context = {'title' : instance.title, 'instance' : instance, 'form' : form}
+    return render(request, 'post_form.html', context)
 
 def post_delete(request, slug=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
     instance=get_object_or_404(Post,slug=slug)
     instance.delete()
     messages.success(request,"<a href='#'>Successfully</a> Deleted !!",extra_tags='html_safe')
